refactor(harness-resources): isolate precedence and snapshot pipeline
